refactor: AutoReshuffleManagerに自動再抽選機能を抽出
- 新規: resources/js/channels/managers/AutoReshuffleManager.js
  - 再生位置監視
  - フェードイン・フェードアウト制御
  - バッファリングタイムアウト検知
  - スタック検知
  - 終了時刻計算
  - 設定の永続化 (sessionStorage)

- archive-list.js: 1177行 → 1001行 (176行削減)
  - 自動再抽選関連メソッドをManagerに委譲
  - toggleAutoReshuffle()をシンプルな委譲メソッドに変更
